requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
rich>=13.0.0       # Pretty terminal output — makes learning more fun
pytest>=8.0.0      # Phase 2 eval framework
